Separate Jedis configured target state
