+++ instrumentation/jedis/jedis-4.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jedis/v4_0/JedisInstrumentationModule.java
@@ -33,7 +33,7 @@ public class JedisInstrumentationModule extends InstrumentationModule
 
   @Override
   public void registerVirtualFields(BiConsumer<String, String> virtualFieldRegistrar) {
-    String configuredTarget = JedisSingletons.ConfiguredTarget.class.getName();
+    String configuredTarget = JedisConfiguredTargets.ConfiguredTarget.class.getName();
     virtualFieldRegistrar.accept(
         "redis.clients.jedis.providers.JedisConnectionProvider", configuredTarget);
     virtualFieldRegistrar.accept(
